refactor Further improved folder structure and delted duplicate functions from old tools
