+++ app/src/components/modal/modal_transaction/index.tsx
@@ -6,7 +6,7 @@ import styled, { withTheme } from 'styled-components'
 import { CONFIRMATION_COUNT } from '../../../common/constants'
 import { useConnectedWeb3Context } from '../../../hooks'
 import { getBlockExplorer, getTxHashBlockExplorerURL } from '../../../util/networks'
-import { isValidHttpUrl } from '../../../util/tools/base'
+import { isValidHttpUrl } from '../../../util/tools'
 import { TransactionStep } from '../../../util/types'
 import { Button } from '../../button'
 import { ButtonType } from '../../button/button_styling_types'
