+++ app/src/components/modal/modal_your_connection/index.tsx
@@ -6,8 +6,7 @@ import styled, { withTheme } from 'styled-components'
 import { useConnectedWeb3Context } from '../../../hooks'
 import { NetworkId, bridgeTokensList, getToken, networkIds } from '../../../util/networks'
 import { getImageUrl } from '../../../util/token'
-import { truncateStringInTheMiddle } from '../../../util/tools'
-import { formatBigNumber, waitForConfirmations } from '../../../util/tools/base'
+import { formatBigNumber, truncateStringInTheMiddle, waitForConfirmations } from '../../../util/tools'
 import { KnownTokenValue, Token, TransactionStep } from '../../../util/types'
 import { Button } from '../../button/button'
 import { ButtonType } from '../../button/button_styling_types'
