+++ app/src/services/xdai.ts
@@ -16,7 +16,7 @@ import {
   XDAI_TO_DAI_TOKEN_BRIDGE_ADDRESS,
 } from '../common/constants'
 import { getInfuraUrl, knownTokens, networkIds } from '../util/networks'
-import { waitABit } from '../util/tools/base'
+import { waitABit } from '../util/tools'
 
 import { ERC20Service } from './erc20'
