+++ app/src/util/tools/fpmm/liquidity/calcDepositedTokens.ts
@@ -1,7 +1,7 @@
 import { BigNumber } from 'ethers/utils'
 
 import { calcRemoveFundingSendAmounts } from './calcRemoveFundingSendAmounts'
-//stays
+//stays check
 /**
  * Compute the amount of collateral that can be obtained via merging after the user
  * removed `removedFunds` of pool shares.
