+++ app/src/util/tools/fpmm/liquidity/calcRemoveFundingSendAmounts.ts
@@ -1,5 +1,5 @@
 import { BigNumber } from 'ethers/utils'
-//stays
+//stays check
 /**
  * Compute the number of outcome tokens that will be sent to the user by the Market Maker
  * after removing `removedFunds` of pool shares.
